requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
httpx==0.26.0
asyncpg==0.29.0
redis==5.0.1
faster-whisper==1.0.0
yt-dlp==2024.1.1
openai==1.10.0
anthropic==0.18.1
pydantic-settings==2.1.0
python-dotenv==1.0.0
